+++ test/test_player_scrabble.py
@@ -2,10 +2,10 @@ import unittest
 from game.models import (Player, BagTiles)
 class TestPlayer(unittest.TestCase):
     def test_init(self):
-        player_1 = Player()
+        player_1 = Player(1)
         self.assertEqual(len(player_1.tiles), 7)
     def test_refreshtiles(self):
-        player_2 = Player()
+        player_2 = Player(2)
         player_2.refreshtiles(2)
         self.assertEqual(len(player_2.tiles), 9)
  #   def test_taketiletilebag(self):