+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-
+# [V 1.1.3] 2023/08/20: Added turn sistem, Added ScrabbleGame class
+### Added a turn sistem in Scrabblegame(class),the game can now have different turns for players
+### Fixed issues with player(class), that didnt allow the class to have and ID number
 # [V 1.1.2] 2023/09/13: Added new class scrabblegame and changed calculatewordvalue method
 ### [V 1.1.2] 2023/09/19: Changed the way the changelog is setup, fixed calculate word value(wip)
 # [V 1.1.1] 2023/09/11: Reworked the Bagtile class